msm: kgsl: increase valid timestamp range
The existing timestamp_cmp function returns a different
result depending on the order of the input parameters due to
having an asymetric valid window. When no rollover is
detected the window is 2^31 but when a rollover is detected
the window is 25000. This change makes the rollover window
symmetric at 2^31.

CRs-Fixed: 347631
Change-Id: I63804b8aa2d51226b506254755f040032048865d
Signed-off-by: Jeff Boody <jboody@codeaurora.org>